+++ spring/src/main/java/com/kartoflane/inteca/eval/spring/rest/FatherNotFoundException.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class FatherNotFoundException extends RuntimeException {
-	FatherNotFoundException(Long id) {
+	FatherNotFoundException(Integer id) {
 		super(String.format("Could not find father with id '%d'.", id));
 	}
 }
